+++ data/src/main/java/com/jesusvilla/data/source/handler/Resource.kt
@@ -1,0 +1,21 @@
+package com.jesusvilla.data.source.handler
+
+
+/**
+ * Created by Jesús Villa on 3/04/22
+ */
+data class Resource<out T>(val status: Status, val data: T?, val message: String?) {
+    companion object {
+        fun <T> success(data: T): Resource<T> {
+            return Resource(Status.SUCCESS, data, null)
+        }
+
+        fun <T> error(msg: String?, data: T?): Resource<T> {
+            return Resource(Status.ERROR, data, msg)
+        }
+
+        fun <T> loading(data: T?): Resource<T> {
+            return Resource(Status.LOADING, data, null)
+        }
+    }
+}