+++ usecases/src/main/java/com/jesusvilla/usecases/GetPopularMovies.kt
@@ -1,8 +1,9 @@
 package com.jesusvilla.usecases
 
 import com.jesusvilla.data.repository.MoviesRepository
+import com.jesusvilla.data.source.handler.Resource
 import com.jesusvilla.domain.Movie
 
 class GetPopularMovies(private val moviesRepository: MoviesRepository) {
-    suspend fun invoke(): List<Movie> = moviesRepository.getPopularMovies()
+    suspend fun invoke(): Resource<List<Movie>> = moviesRepository.getPopularMovies()
 }